* commit '797b5adcb4ead3c687db3d9fd831dff93760dd4a':
  AArch64: Make AssetManager and related classes 64-bit compatible

+++ core/java/android/content/res/Resources.java
@@ -1459,7 +1459,7 @@ public class Resources {
         }
 
         private final AssetManager mAssets;
-        private final int mTheme;
+        private final long mTheme;
     }
 
     /**
